+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- IRS API now requires 'view_irs' resource access inside Keycloak JWT token.
 - New 2.0.0 version of IRS API. Main goal was to remove 'job' prefix from attribute names
     - change 'jobId' to 'id' in GET calls
     - change 'jobState' to 'state' in GET calls